Move mergeBranches inside Merge.
